Fix power factor calculation
According to specifications for the ADE7953, the power factor can
be positive (less than 1 is inductive load) and negative (less
than 1 is capacitive load). The value for a pure resistive load
representing 1 is 0x7fff, everything below this value is a power
factor < 1 and indication for inductive load. Everything above
0x7fff is again less than one, but it is negative and is indication
for capacitive load. In order to achive this - the calculation is
changed. When the value is greater than 0x7fff, eg. 0x800E the
returned value should be -0.999, this is done by -(0x7fff/0x800E)
When the value is below the 0x7fff, eg 0x7f36, the calculation
is 0x7fff/0x7f36. This way the returned value is in the range
of 0 to 1 and with sign representing the type of load connected.